refactored to use Decimal instead of Price

+++ src/matching_engine/engine.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
 use super::orderbook::{Order, Orderbook};
+use rust_decimal::Decimal;
 
 /*
     BTC_USDT
@@ -39,7 +40,7 @@ impl MatchingEngine {
         println!("opening new orderbook: {:?}", pair.to_string());
     }
 
-    pub fn place_limit_order(&mut self, pair: TradingPair, price: f64, order: Order) -> Result<(), String> {
+    pub fn place_limit_order(&mut self, pair: TradingPair, price: Decimal, order: Order) -> Result<(), String> {
         match self.orderbooks.get_mut(&pair) {
             Some(orderbook) => {
                 orderbook.add_order(price, order);
